pipelines/fetch_chess_data/requirements.txt
requests
functions-framework
google-cloud-bigquery
google-cloud-firestore